+++ Libplanet.Extensions.Cocona/Commands/BlockCommand.cs
@@ -7,10 +7,18 @@ using System.Text.Json;
 using Bencodex;
 using global::Cocona;
 using Libplanet.Action;
+using Libplanet.Blockchain;
 using Libplanet.Blocks;
+using Libplanet.Crypto;
 
 public class BlockCommand
 {
+    public enum OutputFormat
+    {
+        Bencodex,
+        Json,
+    }
+
     [Command(Description = "Analyze the given block.")]
     public void Analyze(
         [Argument(
@@ -87,5 +95,61 @@ public class BlockCommand
 
         using var textWriter = new StreamWriter(outputStream);
         textWriter.WriteLine();
+    }
+
+    [Command(Description = "Generate a genesis block.")]
+    public void GenerateGenesis(
+        [Argument(Name = "KEY-ID", Description = "A private key to use for signing.")]
+        Guid keyId,
+        PassphraseParameters passphrase,
+        [Argument(Name = "FILE", Description = "File to write the genesis block to.  " +
+            "Use `-` for stdout (you can still refer to a file named \"-\" by \"./-\").")]
+        string file,
+        BlockPolicyParams blockPolicyParams,
+        [Option('f', Description = "Output format.")]
+        OutputFormat format = OutputFormat.Bencodex
+    )
+    {
+        // FIXME: Declare a ICommandParameterSet type taking key ID and keystore path instead:
+        PrivateKey key = new KeyCommand().UnprotectKey(keyId, passphrase, ignoreStdin: true);
+        Block<NullAction> genesis = BlockChain<NullAction>.MakeGenesisBlock(
+            privateKey: key,
+            blockAction: blockPolicyParams.GetBlockAction(),
+            nativeTokenPredicate: blockPolicyParams is { } p && p.GetNativeTokens() is { } tokens
+                ? tokens.Contains
+                : null
+        );
+        using Stream stream = file == "-"
+            ? Console.OpenStandardOutput()
+            : File.OpenWrite(file);
+        switch (format)
+        {
+            // FIXME: Configure JsonConverter for Block<T>:
+            case OutputFormat.Json:
+                var writerOptions = new JsonWriterOptions { Indented = true };
+                using (var writer = new Utf8JsonWriter(stream, writerOptions))
+                {
+                    var serializerOptions = new JsonSerializerOptions
+                    {
+                        AllowTrailingCommas = false,
+                        DictionaryKeyPolicy = JsonNamingPolicy.CamelCase,
+                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                        IgnoreReadOnlyProperties = false,
+                    };
+                    JsonSerializer.Serialize(writer, genesis, serializerOptions);
+                }
+
+                using (var textWriter = new StreamWriter(stream))
+                {
+                    textWriter.WriteLine();
+                }
+
+                break;
+
+            default:
+                Codec codec = new ();
+                codec.Encode(genesis.MarshalBlock(), stream);
+                break;
+        }
     }
 }